Rajout des checks admins pour les actions sur les produits

+++ controller/ControllerProduit.php
@@ -41,18 +41,22 @@
 
 
 		public static function delete(){
-			if(!is_null(myGet('id'))){
-				$p = ModelProduit::select(myGet('id'));
-				if($p == false){
-					$view='error'; $pagetitle='ErreurProduit'; $errorType = "Delete d'un Produit: id fourni non existant";
-				}else{
-					ModelProduit::delete(myGet('id'));
-					$tab_p = ModelProduit::selectAll();
+			if(Session::is_admin()){
+				if(!is_null(myGet('id'))){
+					$p = ModelProduit::select(myGet('id'));
+					if($p == false){
+						$view='error'; $pagetitle='ErreurProduit'; $errorType = "Delete d'un Produit: id fourni non existant";
+					}else{
+						ModelProduit::delete(myGet('id'));
+						$tab_p = ModelProduit::selectAll();
 
-					$view='delete'; $pagetitle='Suppresion Produit';
+						$view='delete'; $pagetitle='Suppresion Produit';
+					}
+				}else{
+					$view='error'; $pagetitle='ErreurProduit'; $errorType = "Delete d'un Produit: Pas d'id fourni";
 				}
 			}else{
-				$view='error'; $pagetitle='ErreurProduit'; $errorType = "Delete d'un Produit: Pas d'id fourni";
+				$view='error'; $pagetitle='ErreurProduit'; $errorType = "Delete d'un Produit: Acces Restreint";
 			}
 			require (File::build_path(array("view","view.php")));
 		}
@@ -60,46 +64,54 @@
 
 
 		public static function create(){
-    		$pId = "\"\"";
-    		$pNom = "\"\"";
-    		$pDescription = "\"\"";
-    		$pPrix = "\"\"";
-    		$pNationalite = "\"\"";
-    		$pAction = "create";
-		
-			$view='update'; $pagetitle='Creation Produit';
+			if(Session::is_admin()){
+	    		$pId = "\"\"";
+	    		$pNom = "\"\"";
+	    		$pDescription = "\"\"";
+	    		$pPrix = "\"\"";
+	    		$pNationalite = "\"\"";
+	    		$pAction = "create";
+			
+				$view='update'; $pagetitle='Creation Produit';
+			}else{
+				$view='error'; $pagetitle='ErreurProduit'; $errorType = "Create d'un Produit: Acces Restreint";
+			}
 			require (File::build_path(array("view","view.php")));
 		}
 
 
 
 		public static function created(){
-			if (!is_null(myGet('id')) && !is_null(myGet('nom')) && !is_null(myGet('description')) && !is_null(myGet('prix')) && !is_null(myGet('nationalite'))){
+			if(Session::is_admin()){
+				if (!is_null(myGet('id')) && !is_null(myGet('nom')) && !is_null(myGet('description')) && !is_null(myGet('prix')) && !is_null(myGet('nationalite'))){
 
-				$data = array(
-					"id" => myGet('id'),
-					"nom" => myGet('nom'),
-					"description" => myGet('description'),
-					"prix" => myGet('prix'),
-					"nationalite" => myGet('nationalite')
-				);
+					$data = array(
+						"id" => myGet('id'),
+						"nom" => myGet('nom'),
+						"description" => myGet('description'),
+						"prix" => myGet('prix'),
+						"nationalite" => myGet('nationalite')
+					);
 
-				if(!is_null(myGet('pathImgProduit'))){
-					$data["pathImgProduit"] = myGet('pathImgProduit');
+					if(!is_null(myGet('pathImgProduit'))){
+						$data["pathImgProduit"] = myGet('pathImgProduit');
+					}else{
+						$data["pathImgProduit"] = "src/placeholder.jpg";
+					}
+					$p = new ModelProduit($data);
+					var_dump($p->get_object_vars());	
+					if(ModelProduit::save($p) == false){
+						$view='error'; $pagetitle='Erreur de Création'; $errorType = 'Created Produit: id fourni déjà existant';
+						
+					}else{
+						$tab_p = ModelProduit::selectAll();
+						$view='created'; $pagetitle='Création Reussie';
+					}
 				}else{
-					$data["pathImgProduit"] = "src/placeholder.jpg";
-				}
-				$p = new ModelProduit($data);
-				var_dump($p->get_object_vars());	
-				if(ModelProduit::save($p) == false){
-					$view='error'; $pagetitle='Erreur de Création'; $errorType = 'Created Produit: id fourni déjà existant';
-					
-				}else{
-					$tab_p = ModelProduit::selectAll();
-					$view='created'; $pagetitle='Création Reussie';
+					$view='error'; $pagetitle='ErreurProduit'; $errorType = "Create d'un Produit: Problème de paramètres";
 				}
 			}else{
-				$view='error'; $pagetitle='ErreurProduit'; $errorType = "Create d'un Produit: Problème de paramètres";
+				$view='error'; $pagetitle='ErreurProduit'; $errorType = "Create d'un Produit: Acces Restreint";
 			}
 			require (File::build_path(array("view","view.php")));
 		}
@@ -108,27 +120,31 @@
 
 
 		public static function update(){
-			if (!is_null(myGet('id'))){
-				$p = ModelProduit::select(myGet('id'));
+			if(Session::is_admin()){
+				if (!is_null(myGet('id'))){
+					$p = ModelProduit::select(myGet('id'));
 
-				if($p == false){
-					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'update Produit: id fourni non existant';
+					if($p == false){
+						$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'update Produit: id fourni non existant';
+					}
+					else{
+
+	    				$pId = htmlspecialchars($p->get('id'));
+	    				$pNom =htmlspecialchars($p->get('nom'));
+	    				$pDescription = htmlspecialchars($p->get('description'));
+	    				$pPrix = htmlspecialchars($p->get('prix'));
+	    				$pNationalite = htmlspecialchars($p->get('nationalite'));
+	    				$pathImgProduit = $p->get('pathImgProduit');
+	    				$pAction = "update";
+
+						$view='update'; $pagetitle='Mise A Jour';
+					}
 				}
 				else{
-
-    				$pId = htmlspecialchars($p->get('id'));
-    				$pNom =htmlspecialchars($p->get('nom'));
-    				$pDescription = htmlspecialchars($p->get('description'));
-    				$pPrix = htmlspecialchars($p->get('prix'));
-    				$pNationalite = htmlspecialchars($p->get('nationalite'));
-    				$pathImgProduit = $p->get('pathImgProduit');
-    				$pAction = "update";
-
-					$view='update'; $pagetitle='Mise A Jour';
+					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'update Produit: Problème de paramètres';
 				}
-			}
-			else{
-				$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'update Produit: Problème de paramètres';
+			}else{
+				$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'update Produit: Acces Restreint';
 			}
 			require (File::build_path(array("view","view.php")));
 		}
@@ -136,48 +152,56 @@
 
 
 		public static function updated(){
-			if (!is_null(myGet('id')) && !is_null(myGet('nom')) && !is_null(myGet('description')) && !is_null(myGet('prix')) && !is_null(myGet('nationalite'))){
-				$data = array(
-					"id" => myGet('id'),
-					"nom" => myGet('nom'),
-					"description" => myGet('description'),
-					"prix" => myGet('prix'),
-					"nationalite" => myGet('nationalite'),
-					"pathImgProduit" => myGet('pathImgProduit'),
-				);
-				if (ModelProduit::select($_GET['id']) != false) {
-					if(ModelProduit::update($data)){
-						$tab_p = ModelProduit::selectAll();
-						$view='updated'; $pagetitle='Mise A Jour';
+			if(Session::is_admin()){
+				if (!is_null(myGet('id')) && !is_null(myGet('nom')) && !is_null(myGet('description')) && !is_null(myGet('prix')) && !is_null(myGet('nationalite'))){
+					$data = array(
+						"id" => myGet('id'),
+						"nom" => myGet('nom'),
+						"description" => myGet('description'),
+						"prix" => myGet('prix'),
+						"nationalite" => myGet('nationalite'),
+						"pathImgProduit" => myGet('pathImgProduit'),
+					);
+					if (ModelProduit::select($_GET['id']) != false) {
+						if(ModelProduit::update($data)){
+							$tab_p = ModelProduit::selectAll();
+							$view='updated'; $pagetitle='Mise A Jour';
+						}else{
+							$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'updated Produit: Probleme rencontré lors de la maj';
+						}
 					}else{
-						$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'updated Produit: Probleme rencontré lors de la maj';
+						$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'updated Produit: id Produit non existant';
 					}
-				}else{
-					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'updated Produit: id Produit non existant';
 				}
-			}
-			else{
-				$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'updated Produit: Problème de paramètres';
+				else{
+					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'updated Produit: Problème de paramètres';
+				}
+			}else{
+				$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'updated Produit: Acces Restreint';
 			}
 			require (File::build_path(array("view","view.php")));
 		}
 
 
 		public static function imgupload(){
-			if (!is_null(myGet('id'))){
-				$p = ModelProduit::select(myGet('id'));
+			if(Session::is_admin()){
+				if (!is_null(myGet('id'))){
+					$p = ModelProduit::select(myGet('id'));
 
-				if($p == false){
-					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgupload Produit: id fourni non existant';
-					
-				}
-				else{
-					$pId = htmlspecialchars($p->get('id'));
-					$view='imgupload'; $pagetitle='Upload Img Produit';
-				}
+					if($p == false){
+						$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgupload Produit: id fourni non existant';
+						
+					}
+					else{
+						$pId = htmlspecialchars($p->get('id'));
+						$view='imgupload'; $pagetitle='Upload Img Produit';
+					}
 
+				}else{
+					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgupload Produit: Problème de paramètres';
+				}
 			}else{
-				$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgupload Produit: Problème de paramètres';
+					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgupload Produit: Acces Restreint';
 			}
 			require (File::build_path(array("view","view.php")));
 		}
@@ -185,47 +209,51 @@
 
 
 		public static function imguploaded(){
-			if (!is_null(myGet('id'))){
-				$p = ModelProduit::select(myGet('id'));
-				if($p == false){
-					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imguploaded Produit: id fourni non existant';
-					
-				}
-				else{
-					if (!empty($_FILES['nom-du-fichier']) && is_uploaded_file($_FILES['nom-du-fichier']['tmp_name'])) {
-						$name = $_FILES['nom-du-fichier']['name'];
-						$pic_path =  File::build_path(array("src",$name));
+			if(Session::is_admin()){
+				if (!is_null(myGet('id'))){
+					$p = ModelProduit::select(myGet('id'));
+					if($p == false){
+						$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imguploaded Produit: id fourni non existant';
+						
+					}
+					else{
+						if (!empty($_FILES['nom-du-fichier']) && is_uploaded_file($_FILES['nom-du-fichier']['tmp_name'])) {
+							$name = $_FILES['nom-du-fichier']['name'];
+							$pic_path =  File::build_path(array("src",$name));
 
-						$allowed_ext = array("jpg", "jpeg", "png");
-						$explode = explode('.',$_FILES['nom-du-fichier']['name']);
-						if (!in_array(end($explode), $allowed_ext)) {
- 						 	$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'addedimg Produit: Mauvais type de fichier';
-						}else{
-							if (!move_uploaded_file($_FILES['nom-du-fichier']['tmp_name'], $pic_path)) {
-							  	$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imguploaded Produit: La Copie a échouée';
+							$allowed_ext = array("jpg", "jpeg", "png");
+							$explode = explode('.',$_FILES['nom-du-fichier']['name']);
+							if (!in_array(end($explode), $allowed_ext)) {
+	 						 	$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'addedimg Produit: Mauvais type de fichier';
 							}else{
-								$p->set("pathImgProduit","src/$name");
-								
-								if(ModelProduit::update($p->get_object_vars()) == true){
-									$pId = htmlspecialchars($p->get('id'));
-									$path = $p->get('pathImgProduit');
-									$view='imguploaded'; $pagetitle='Uploaded Img Produit';
-				
-									}else{
-										$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'addedimg Produit: ce produit a déjà une image';
-									}
-	
+								if (!move_uploaded_file($_FILES['nom-du-fichier']['tmp_name'], $pic_path)) {
+								  	$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imguploaded Produit: La Copie a échouée';
+								}else{
+									$p->set("pathImgProduit","src/$name");
+									
+									if(ModelProduit::update($p->get_object_vars()) == true){
+										$pId = htmlspecialchars($p->get('id'));
+										$path = $p->get('pathImgProduit');
+										$view='imguploaded'; $pagetitle='Uploaded Img Produit';
+					
+										}else{
+											$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'addedimg Produit: ce produit a déjà une image';
+										}
+		
+								}
 							}
+
+						}else{
+							$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'addedimg Produit: Problème de paramètre fichier';
 						}
 
-					}else{
-						$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'addedimg Produit: Problème de paramètre fichier';
 					}
 
+				}else{
+					$view='error'; $pagetitle='Erreur MAJ'; $errorType = "addedimg Produit: Problème de paramètre d'id";
 				}
-
 			}else{
-				$view='error'; $pagetitle='Erreur MAJ'; $errorType = "addedimg Produit: Problème de paramètre d'id";
+				$view='error'; $pagetitle='Erreur MAJ'; $errorType = "addedimg Produit: Acces Restreint";
 			}
 			require (File::build_path(array("view","view.php")));
 		}
@@ -233,29 +261,33 @@
 
 
 		public static function imgdelete(){
-			if (!is_null(myGet('id'))){
-				$p = ModelProduit::select(myGet('id'));
+			if(Session::is_admin()){
+				if (!is_null(myGet('id'))){
+					$p = ModelProduit::select(myGet('id'));
 
-				if($p == false){
-					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgdelete Produit: id fourni non existant';
-					
-				}
-				else{
-					$pic = $p->get('pathImgProduit');
-					if ($pic != NULL) {
-						$p->set('pathImgProduit',NULL);
-						ModelProduit::update($p->get_object_vars());
-						unlink($pic);
-						$path=false;
-						$view='imgdelete'; $pagetitle='Upload Img Produit';
+					if($p == false){
+						$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgdelete Produit: id fourni non existant';
+						
 					}
 					else{
-						$view='error'; $pagetitle='Erreur MAJ'; $errorType = "imgdelete Produit:Le produit n'a pas d'image";
+						$pic = $p->get('pathImgProduit');
+						if ($pic != NULL) {
+							$p->set('pathImgProduit',NULL);
+							ModelProduit::update($p->get_object_vars());
+							unlink($pic);
+							$path=false;
+							$view='imgdelete'; $pagetitle='Upload Img Produit';
+						}
+						else{
+							$view='error'; $pagetitle='Erreur MAJ'; $errorType = "imgdelete Produit:Le produit n'a pas d'image";
+						}
 					}
-				}
 
+				}else{
+					$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgdelete Produit: Problème de paramètres';
+				}
 			}else{
-				$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgdelete Produit: Problème de paramètres';
+				$view='error'; $pagetitle='Erreur MAJ'; $errorType = 'imgdelete Produit: Acces Restreint';
 			}
 			require (File::build_path(array("view","view.php")));
 		}
